+++ packages/cook/src/cook.ts
@@ -14,6 +14,7 @@ import type {
   LVal,
   NewExpression,
   ObjectPattern,
+  Pattern,
   PatternLike,
   RestElement,
   Statement,
@@ -1534,6 +1535,7 @@ export function cook(
 
   // https://tc39.es/ecma262/#sec-runtime-semantics-namedevaluation
   function NamedEvaluation(node: FunctionDefinition, name: string) {
+    hooks.beforeEvaluate?.(node);
     // No ParenthesizedExpression in ESTree.
     switch (node.type) {
       case "FunctionExpression":
@@ -1806,6 +1808,13 @@ export function cook(
         value: lexicalThis ? Mode.LEXICAL : Mode.STRICT,
       },
     });
+
+    const len = ExpectedArgumentCount(sourceNode.params);
+    Object.defineProperty(F, "length", {
+      configurable: true,
+      value: len,
+    });
+
     if (debug || externalSourceForDebug) {
       Object.defineProperty(F, DebuggerCall, {
         value: function () {
@@ -1815,6 +1824,22 @@ export function cook(
       });
     }
     return F;
+  }
+
+  function ExpectedArgumentCount(
+    params: (Identifier | Pattern | RestElement)[]
+  ) {
+    let count = 0;
+    for (const param of params) {
+      switch (param.type) {
+        case "AssignmentPattern":
+        case "RestElement":
+          return count;
+        default:
+          count++;
+      }
+    }
+    return count;
   }
 
   // Patterns initialization.