+++ packages/brick-playground/CHANGELOG.md
@@ -3,6 +3,84 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.14.11](https://github.com/easyops-cn/next-core/compare/@next-core/brick-playground@1.14.10...@next-core/brick-playground@1.14.11) (2024-07-02)
+
+**Note:** Version bump only for package @next-core/brick-playground
+
+
+
+
+
+## [1.14.10](https://github.com/easyops-cn/next-core/compare/@next-core/brick-playground@1.14.9...@next-core/brick-playground@1.14.10) (2024-06-26)
+
+**Note:** Version bump only for package @next-core/brick-playground
+
+
+
+
+
+## [1.14.9](https://github.com/easyops-cn/next-core/compare/@next-core/brick-playground@1.14.8...@next-core/brick-playground@1.14.9) (2024-06-26)
+
+**Note:** Version bump only for package @next-core/brick-playground
+
+
+
+
+
+## [1.14.8](https://github.com/easyops-cn/next-core/compare/@next-core/brick-playground@1.14.7...@next-core/brick-playground@1.14.8) (2024-06-25)
+
+**Note:** Version bump only for package @next-core/brick-playground
+
+
+
+
+
+## [1.14.7](https://github.com/easyops-cn/next-core/compare/@next-core/brick-playground@1.14.6...@next-core/brick-playground@1.14.7) (2024-06-24)
+
+
+### Bug Fixes
+
+* respect gap settings of examples in playground ([d090b4e](https://github.com/easyops-cn/next-core/commit/d090b4eb18d53b713442d828b9873821cccb3efa))
+
+
+
+
+
+## [1.14.6](https://github.com/easyops-cn/next-core/compare/@next-core/brick-playground@1.14.5...@next-core/brick-playground@1.14.6) (2024-06-24)
+
+**Note:** Version bump only for package @next-core/brick-playground
+
+
+
+
+
+## [1.14.5](https://github.com/easyops-cn/next-core/compare/@next-core/brick-playground@1.14.4...@next-core/brick-playground@1.14.5) (2024-06-21)
+
+**Note:** Version bump only for package @next-core/brick-playground
+
+
+
+
+
+## [1.14.4](https://github.com/easyops-cn/next-core/compare/@next-core/brick-playground@1.14.3...@next-core/brick-playground@1.14.4) (2024-06-21)
+
+**Note:** Version bump only for package @next-core/brick-playground
+
+
+
+
+
+## [1.14.3](https://github.com/easyops-cn/next-core/compare/@next-core/brick-playground@1.14.2...@next-core/brick-playground@1.14.3) (2024-06-21)
+
+
+### Bug Fixes
+
+* inject local editors in misc ([ac58c89](https://github.com/easyops-cn/next-core/commit/ac58c898676f8600b6e8443f362a672d5b494248))
+
+
+
+
+
 ## [1.14.2](https://github.com/easyops-cn/next-core/compare/@next-core/brick-playground@1.14.1...@next-core/brick-playground@1.14.2) (2024-06-21)
 
 **Note:** Version bump only for package @next-core/brick-playground