+++ packages/easyops-runtime/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.10.27](https://github.com/easyops-cn/next-core/compare/@next-core/easyops-runtime@0.10.26...@next-core/easyops-runtime@0.10.27) (2024-07-02)
+
+**Note:** Version bump only for package @next-core/easyops-runtime
+
+
+
+
+
+## [0.10.26](https://github.com/easyops-cn/next-core/compare/@next-core/easyops-runtime@0.10.25...@next-core/easyops-runtime@0.10.26) (2024-06-26)
+
+**Note:** Version bump only for package @next-core/easyops-runtime
+
+
+
+
+
+## [0.10.25](https://github.com/easyops-cn/next-core/compare/@next-core/easyops-runtime@0.10.24...@next-core/easyops-runtime@0.10.25) (2024-06-26)
+
+**Note:** Version bump only for package @next-core/easyops-runtime
+
+
+
+
+
+## [0.10.24](https://github.com/easyops-cn/next-core/compare/@next-core/easyops-runtime@0.10.23...@next-core/easyops-runtime@0.10.24) (2024-06-25)
+
+**Note:** Version bump only for package @next-core/easyops-runtime
+
+
+
+
+
+## [0.10.23](https://github.com/easyops-cn/next-core/compare/@next-core/easyops-runtime@0.10.22...@next-core/easyops-runtime@0.10.23) (2024-06-24)
+
+**Note:** Version bump only for package @next-core/easyops-runtime
+
+
+
+
+
+## [0.10.22](https://github.com/easyops-cn/next-core/compare/@next-core/easyops-runtime@0.10.21...@next-core/easyops-runtime@0.10.22) (2024-06-21)
+
+**Note:** Version bump only for package @next-core/easyops-runtime
+
+
+
+
+
+## [0.10.21](https://github.com/easyops-cn/next-core/compare/@next-core/easyops-runtime@0.10.20...@next-core/easyops-runtime@0.10.21) (2024-06-21)
+
+**Note:** Version bump only for package @next-core/easyops-runtime
+
+
+
+
+
 ## [0.10.20](https://github.com/easyops-cn/next-core/compare/@next-core/easyops-runtime@0.10.19...@next-core/easyops-runtime@0.10.20) (2024-06-21)