+++ packages/eslint-config-next/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.22](https://github.com/easyops-cn/next-core/compare/@next-core/eslint-config-next@2.0.21...@next-core/eslint-config-next@2.0.22) (2024-06-21)
+
+**Note:** Version bump only for package @next-core/eslint-config-next
+
+
+
+
+
 ## [2.0.21](https://github.com/easyops-cn/next-core/compare/@next-core/eslint-config-next@2.0.20...@next-core/eslint-config-next@2.0.21) (2024-06-06)